update docstrings and move protos
