+++ chemprop/v2/featurizers/__init__.py
@@ -1,6 +1,6 @@
-from .proto import MolGraphFeaturizerProto
-from .atom import AtomFeaturizerProto, AtomFeaturizer
-from .bond import BondFeaturizerProto, BondFeaturizer
+from .protos import *
+from .atom import AtomFeaturizer
+from .bond import BondFeaturizer
 from .molgraph import MolGraph, BatchMolGraph
-from .molecule import MoleculeMolGraphFeaturizerProto, MoleculeMolGraphFeaturizer
-from .reaction import ReactionMolGraphFeaturizerProto, ReactionMolGraphFeaturizer, ReactionMode
+from .molecule import MoleculeMolGraphFeaturizer
+from .reaction import RxnMolGraphFeaturizer, RxnMode
