+++ chemprop/v2/featurizers/utils.py
@@ -1,9 +1,7 @@
-from enum import auto
 from typing import Sequence
 
 import numpy as np
 
-from chemprop.v2.utils import AutoName
 
 
 class MultiHotFeaturizerMixin:
@@ -23,25 +21,3 @@ class MultiHotFeaturizerMixin:
         n = len(xs)
 
         return xs.index(x) if x in xs else n, n + 1
-
-
-class ReactionMode(AutoName):
-    """The manner in which a reaction should be featurized into a `MolGraph`"""
-
-    REAC_PROD = auto()
-    """concatenate the reactant features with the product features."""
-    REAC_PROD_BALANCE = auto()
-    """concatenate the reactant features with the products feature and balances imbalanced
-    reactions"""
-    REAC_DIFF = auto()
-    """concatenates the reactant features with the difference in features between reactants and
-    products"""
-    REAC_DIFF_BALANCE = auto()
-    """concatenates the reactant features with the difference in features between reactants and
-    product and balances imbalanced reactions"""
-    PROD_DIFF = auto()
-    """concatenates the product features with the difference in features between reactants and
-    products"""
-    PROD_DIFF_BALANCE = auto()
-    """concatenates the product features with the difference in features between reactants and
-    products and balances imbalanced reactions"""
